pyproject.toml
[tool.poetry]
name = "sptest"
version = "0.1.0"
description = ""
authors = ["faruryo <faruryo@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.9"
typer = "^0.3.2"
speedtest-cli = "^2.1.2"
ambient = {git = "https://github.com/AmbientDataInc/ambient-python-lib.git"}
pandas = "1.1.5"
numpy = "1.18.5"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
flake8 = "^3.8.4"
black = "^20.8b1"
isort = "^5.6.4"
mypy = "^0.790"
rope = "^0.18.0"
pytest-mock = "^3.4.0"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
